+++ docs/src/app/pages/docs/components/app-components-charts/app-components-charts.tsx
@@ -63,7 +63,7 @@ export class AppComponentsCharts {
 <li><a href="#app-components-chart-animated-bars">Animated bars</a></li>
 </ul>
 </li>
-<li><a href="#app-components-chart-getting-started">Getting started</a><ul>
+<li><a href="#app-components-chart-installation">Installation</a><ul>
 <li><a href="#app-components-chart-using-deckdeckgo-charts-from-a-cdn">Using DeckDeckGo charts from a CDN</a></li>
 <li><a href="#app-components-chart-install-deckdeckgo-charts-from-npm">Install DeckDeckGo charts from NPM</a></li>
 <li><a href="#app-components-chart-framework-integration">Framework integration</a></li>
@@ -174,10 +174,11 @@ export class AppComponentsCharts {
   </div>
 </div>
 
-<h2 id="app-components-charts-getting-started">Getting started</h2>
-<p>To create easily your PWA presentation and to enjoy all the options, I suggest you to create your slides using the CLI as described in the <a href="/docs/introduction">Getting started chapter</a>.</p>
-<p>Doing so you will use the <a href="https://deckdeckgo.com">DeckDeckGo</a> starter kit which already includes this Web Component.</p>
-<p>However, if you are looking to use this Web Component as a standalone component, to add a charts to your web applications, it could be use directly in your project from a CDN, using a simple script include, or could be installed from <a href="https://www.npmjs.com/package/@deckdeckgo/charts">npm</a>.</p>
+<h2 id="app-components-charts-installation">Installation</h2>
+<p>This component could be added to your web application using the following methods.</p>
+<blockquote>
+<p>If you are using our Starter Kit to develop your presentation, no need to worry about this, this component is included, therefore you could skip the &quot;Installation&quot; chapter.</p>
+</blockquote>
 <h3 id="app-components-charts-using-deckdeckgo-charts-from-a-cdn">Using DeckDeckGo charts from a CDN</h3>
 <p>It&#39;s recommended to use <a href="https://unpkg.com/">unpkg</a> to use the <a href="https://deckdeckgo.com">DeckDeckGo</a> charts from a CDN. To do so, add the following include script in the main HTML file of your project:</p>
 <deckgo-highlight-code language="javascript">
@@ -188,7 +189,14 @@ export class AppComponentsCharts {
       <code slot="code">npm install @deckdeckgo&#47;charts</code>
     </deckgo-highlight-code><h3 id="app-components-charts-framework-integration">Framework integration</h3>
 <p>The <a href="https://stenciljs.com/docs/overview">Stencil documentation</a> provide examples of framework integration for <a href="https://stenciljs.com/docs/angular">Angular</a>, <a href="https://stenciljs.com/docs/react">React</a>, <a href="https://stenciljs.com/docs/vue">Vue</a> and <a href="https://stenciljs.com/docs/ember">Ember</a>.</p>
-<h2 id="app-components-charts-usage">Usage</h2>
+<p>That being said, commonly, you might either <code>import</code> or <code>load</code> it:</p>
+<h4 id="app-components-charts-import">Import</h4>
+<deckgo-highlight-code language="javascript">
+      <code slot="code">import &#039;@deckdeckgo&#47;charts&#039;;</code>
+    </deckgo-highlight-code><h4 id="app-components-charts-loader">Loader</h4>
+<deckgo-highlight-code language="javascript">
+      <code slot="code">import &#123; defineCustomElements as deckDeckGoElement &#125; from &#039;@deckdeckgo&#47;charts&#47;dist&#47;loader&#039;;{'\n'}deckDeckGoElement(window);</code>
+    </deckgo-highlight-code><h2 id="app-components-charts-usage">Usage</h2>
 <p>The main idea behind the <a href="https://deckdeckgo.com">DeckDeckGo</a> charts is to be able to drop a <strong>csv</strong> file, provide it to the component and let it do the job respectively draw the chart.</p>
 <h3 id="app-components-charts-pie-usage">Pie usage</h3>
 <p>The <code>&lt;deckgo-pie-chart/&gt;</code> Web Component draw a Pie chart.</p>