+++ docs/src/app/pages/docs/components/app-components-gif/app-components-gif.tsx
@@ -53,6 +53,7 @@ export class AppComponentsGif {
 <h2 id="app-components-gif-table-of-contents">Table of contents</h2>
 <ul>
 <li><a href="#app-components-gif-showcase">Showcase</a></li>
+<li><a href="#app-components-gif-installation">Installation</a></li>
 <li><a href="#app-components-gif-usage">Usage</a><ul>
 <li><a href="#app-components-gif-slots">Slots</a></li>
 </ul>
@@ -69,6 +70,11 @@ export class AppComponentsGif {
   </deckgo-gif>
 </div>
 
+<h2 id="app-components-gif-installation">Installation</h2>
+<p>This component is part of the &quot;Gif&quot; template. Therefore, if you would like to use it, install the related slide as described in its <a href="/slides/gif">installation</a> chapter.</p>
+<blockquote>
+<p>If you are using our Starter Kit to develop your presentation, no need to worry about this, this component is included, therefore you could skip the &quot;Installation&quot; chapter.</p>
+</blockquote>
 <h2 id="app-components-gif-usage">Usage</h2>
 <p>The &quot;Gif&quot; slide&#39;s Web Component could be integrated using the tag <code>&lt;deckgo-gif/&gt;</code>.</p>
 <deckgo-highlight-code language="javascript">