+++ webcomponents/slides/utils/src/interfaces/deckdeckgo-slide.ts
@@ -1,0 +1,11 @@
+export interface DeckdeckgoSlide {
+  beforeSwipe(enter: boolean, _reveal: boolean): Promise<boolean>;
+
+  afterSwipe(): Promise<void>;
+
+  lazyLoadContent(): Promise<void>;
+
+  revealContent(): Promise<void>;
+
+  hideContent(): Promise<void>;
+}